+++ src/app/app.routes.js
@@ -1,0 +1,12 @@
+import { HomeComponent, PageNotFoundComponent } from './components';
+
+export const routes = [
+  {
+    path: '/',
+    component: HomeComponent
+  },
+  {
+    path: '*',
+    component: PageNotFoundComponent
+  }
+];