+++ src/app/components/page-not-found/page-not-found.component.jsx
@@ -3,7 +3,14 @@ import React from 'react';
 
 import './page-not-found.component.scss';
 
+/**
+ * Page not found component.
+ * @extends React.Component
+ */
 export class PageNotFoundComponent extends React.Component {
+  /**
+   * Render method.
+   */
   render() {
     return (
       <div>