+++ src/vendor.js
@@ -5,8 +5,8 @@ import 'react-dom';
 import 'react-router';
 
 // Redux
-import 'redux';
+/*import 'redux';
 import 'react-redux';
-import 'react-router-redux';
+import 'react-router-redux';*/
 
 // other vendors here...