+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This project adheres to [Semantic Versioning](http://semver.org).
 ### Added
 - Added jsdoc support
 
+### Changed
+- Upgraded to react-router@4.0.0-alpha.6
+
+### Fixed
+- Routes now work properly with HMR
+
 ## [0.3.1] - 2016-11-22
 ### Added
 - Added flow support